+++ src/main/java/pl/kupniewski/application/order/CreateOrderByCustomer.java
@@ -3,6 +3,7 @@ package pl.kupniewski.application.order;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Component;
 import pl.kupniewski.application.order.Order.Status;
+import pl.kupniewski.application.storages.OrderStorage;
 
 @AllArgsConstructor
 @Component
