+++ src/main/java/pl/kupniewski/application/restaurant/CompleteOrderByRestaurant.java
@@ -1,9 +1,10 @@
-package pl.kupniewski.application.deliver;
+package pl.kupniewski.application.restaurant;
 
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Component;
-import pl.kupniewski.application.deliver.api.DelivererApi;
+import pl.kupniewski.application.deliver.OrderReadyToDeliver;
 import pl.kupniewski.application.order.Order;
+import pl.kupniewski.application.storages.DelivererStorage;
 
 @Component
 @AllArgsConstructor
